requirements.txt
langgraph>=0.2.0
langchain>=0.2.0
pydantic>=2.7.0
absl-py==2.2.2
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aioice==0.10.1
aiortc==1.11.0
aiosignal==1.3.2
aiosqlite==0.21.0
amqp==5.3.1
annotated-types==0.7.0
anyascii==0.3.2
anyio==4.9.0
asttokens==3.0.0
async-timeout==4.0.3
attrs==25.3.0
audioread==3.0.1
av==14.3.0
babel==2.17.0
bangla==0.0.2
bcrypt==4.3.0
billiard==4.2.1
blinker==1.9.0
blis==1.2.1
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.7
boto3==1.38.30
botocore==1.38.30
catalogue==2.0.10
celery==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cloudpathlib==0.21.0
cloudpickle==3.1.1
coloredlogs==15.0.1
confection==0.1.5
contourpy==1.2.1
coqpit==0.0.17
cryptography==44.0.2
ctranslate2==4.6.0
cycler==0.12.1
cymem==2.0.11
Cython==3.0.12
dask==2025.9.1
dateparser==1.1.8
decorator==5.2.1
diskcache==5.6.3
distro==1.9.0
dnspython==2.7.0
docker==7.1.0
docopt==0.6.2
einops==0.8.1
encodec==0.1.1
exceptiongroup==1.2.2
executing==2.2.0
faiss-cpu==1.10.0
fastapi==0.115.12
faster-whisper==1.1.1
filelock==3.18.0
Flask==3.1.0
flatbuffers==25.2.10
fonttools==4.57.0
frozenlist==1.6.0
fsspec==2025.3.2
g2pkk==0.1.2
google-crc32c==1.7.1
greenlet==3.2.0
grpcio==1.71.0
gruut==2.2.3
gruut-ipa==0.13.0
gruut_lang_de==2.0.1
gruut_lang_en==2.0.1
gruut_lang_es==2.0.1
gruut_lang_fr==2.0.2
gTTS==2.5.4
h11==0.14.0
hangul-romanize==0.1.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanfriendly==10.0
idna==3.10
ifaddr==0.2.0
importlib_metadata==8.7.0
inflect==7.5.0
iniconfig==2.1.0
ipdb==0.13.13
ipython==8.35.0
itsdangerous==2.2.0
jamo==0.4.1
jedi==0.19.2
jieba==0.42.1
Jinja2==3.1.6
jiter==0.9.0
jmespath==1.0.1
joblib==1.4.2
jsonlines==1.2.0
jsonpatch==1.33
jsonpointer==3.0.0
jwt==1.3.1
kiwisolver==1.4.8
kombu==5.5.3
langchain==0.3.23
langchain-core==0.3.54
langchain-text-splitters==0.3.8
langcodes==3.5.0
langsmith==0.3.32
language_data==1.3.0
lazy_loader==0.4
librosa==0.10.0
llama_cpp_python==0.3.8
llvmlite==0.43.0
locket==1.0.0
marisa-trie==1.2.1
Markdown==3.8
markdown-it-py==3.0.0
MarkupSafe==3.0.2
matplotlib==3.8.4
matplotlib-inline==0.1.7
mdurl==0.1.2
more-itertools==10.6.0
mpmath==1.3.0
msgpack==1.1.0
multidict==6.4.3
murmurhash==1.0.12
networkx==2.8.8
nltk==3.9.1
noisereduce==3.0.3
num2words==0.5.14
numba==0.60.0
numpy==1.26.4
onnxruntime==1.21.1
openai==1.75.0
openai-whisper==20240930
orjson==3.10.16
packaging==24.2
pandas==2.3.2
parso==0.8.4
partd==1.4.2
pexpect==4.9.0
pillow==11.2.1
platformdirs==4.3.7
pluggy==1.5.0
pooch==1.8.2
preshed==3.0.9
prompt_toolkit==3.0.51
propcache==0.3.1
protobuf==6.30.2
psutil==7.0.0
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==21.0.0
pycparser==2.22
pydantic==2.11.3
pydantic_core==2.33.1
pydub==0.25.1
pyee==13.0.0
pygame==2.6.1
Pygments==2.19.1
PyJWT==2.10.1
pylibsrtp==0.12.0
pynndescent==0.5.13
pyOpenSSL==25.0.0
pyparsing==3.2.3
pypinyin==0.54.0
pysbd==0.3.4
pytest==8.3.5
python-crfsuite==0.9.11
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
python-multipart==0.0.20
pytz==2025.2
PyYAML==6.0.2
redis==6.0.0
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
rich==14.0.0
s3transfer==0.13.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.11.4
sentence-transformers==4.1.0
shellingham==1.5.4
six==1.17.0
smart-open==7.1.0
sniffio==1.3.1
soundfile==0.13.1
soxr==0.5.0.post1
spacy==3.8.5
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.40
srsly==2.5.1
stack-data==0.6.3
starlette==0.46.2
SudachiDict-core==20250129
SudachiPy==0.6.10
sympy==1.13.3
tenacity==9.1.2
tensorboard==2.19.0
tensorboard-data-server==0.7.2
thinc==8.3.4
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
tomli==2.2.1
toolz==1.0.0
torch==2.7.0
torchaudio==2.6.0
tqdm==4.67.1
trainer==0.0.36
traitlets==5.14.3
transformers==4.51.3
typeguard==4.4.2
typer==0.15.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
umap-learn==0.5.7
Unidecode==1.3.8
urllib3==2.4.0
uvicorn==0.34.1
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.13
weasel==0.4.1
webrtcvad==2.0.10
websockets==15.0.1
Werkzeug==3.1.3
wrapt==1.17.2
yarl==1.20.0
zipp==3.23.0
zstandard==0.23.0
